More OOP Part 2
Fixed and finished the remainder of More OOP. To merge branches.
